% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts.R
\name{spell_event}
\alias{spell_event}
\title{Mark changes in spells as events}
\usage{
spell_event(x, right_censored = TRUE)
}
\arguments{
\item{x}{a time ordered vector with values identifying a spell. It is
assumed that when a value in this vector changes that the spell has ended.}

\item{right_censored}{logical. If \code{TRUE} the final observation is treated
as right-censored, i.e. coded as \code{0}. If \code{FALSE} the the final
observation is treated as an event.}
}
\value{
a vector of event codes for each observation in \code{x}.
\code{0} indicates that there was no spell change. \code{1} indicates a
spell change.
}
\description{
Mark changes in spells as events
}
\examples{
x <- c(rep('a', 4), rep('b', 3), 'c', rep('a', 2), 'c')
spell_event(x)
spell_event(x, right_censored = FALSE)

}

