% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_feartures.R
\name{spell_features}
\alias{spell_features}
\title{Find features of a spell that are useful for estimating survival models
with interval censored observations}
\usage{
spell_features(df, var, group, right_censored = TRUE)
}
\arguments{
\item{df}{data frame with the spell variable}

\item{var}{character string specifying the variable in \code{df} with
information specifying the spell.}

\item{group}{character string specifying a group variable (if any) to find
spell features within groups.}

\item{right_censored}{logical. If \code{TRUE} the final observation is treated
as right-censored, i.e. coded as \code{0}. If \code{FALSE} the the final
observation is treated as an event.}
}
\value{
Four columns in addition to the original data frame of the
\code{spell_id}, a binary \code{spell_event} variable, as well as
\code{spell_time1} and \code{spell_time2} with the interval start and
end times within each spell.
}
\description{
Find features of a spell that are useful for estimating survival models
with interval censored observations
}
\details{
Note: assumes that \code{var} is in time order and that observations
are equally spaced in time.
}
\examples{
# Create fake data
group_id <- c(rep('aa', 7), rep('bb', 4))
x <- c(rep('a', 4), rep('b', 3), 'c', rep('a', 2), 'c')
fake_data <- data.frame(group_id = group_id, spell = x)

# Find spell features
spell_features(fake_data, var = 'spell')
spell_features(fake_data, var = 'spell', group = 'group_id')
spell_features(fake_data, var = 'spell', group = 'group_id',
               right_censored = FALSE)

}

