% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts.R
\name{spell_id}
\alias{spell_id}
\title{Find unique spell IDs}
\usage{
spell_id(x)
}
\arguments{
\item{x}{a time ordered vector with values identifying a spell. It is
assumed that when a value in this vector changes that the spell has ended.}
}
\value{
a vector of spell IDs for each observation in \code{x}.
}
\description{
Find unique spell IDs
}
\examples{
x <- c(rep('a', 4), rep('b', 3), 'c', rep('a', 2), 'c')
spell_id(x)

}

