% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts.R
\name{spell_new}
\alias{spell_new}
\title{Mark new spell as events}
\usage{
spell_new(x)
}
\arguments{
\item{x}{a time ordered vector with values identifying a spell. It is
assumed that when a value in this vector changes that the spell has ended.}
}
\value{
a vector of event codes for each observation in \code{x}.
\code{0} indicates that there was no spell change. \code{1} indicates a new
spell.
}
\description{
Mark new spell as events
}
\examples{
x <- c(rep('a', 4), rep('b', 3), 'c', rep('a', 2), 'c')
spell_new(x)

}

